Support nested objects and keywords as first item
Before it was possible to write an array that would result in the
fetching of references or the calling of a function. Now the first item
in the array is quoted whenever it is a reserved sendscript keyword.

SendScript now also supports not just nested arrays but also nested
objects. This is because of the change in implementation that now uses
the JSON revivor and replacer callbacks.